Fix MKN computation for varint-encoded n-grams
The MKN (Modified Kneser-Ney) continuation count computation was broken
for vocabulary-indexed n-grams stored as LEB128 varint-encoded Latin-1
strings:

- Sharded mode used `ngram.split('|')` expecting pipe-delimited format
- Single-trie mode used `ngram.split_whitespace()` which incorrectly
  split when varint bytes contained whitespace (e.g., index 32 = space)

This caused false positives: 75,096 "continuation pairs" from 5.8M
unigrams were actually n-grams whose varint encoding contained
whitespace bytes.

Changes:
- Add `encode_indices_to_key()` helper for encoding context sub-sequences
- Fix `compute_continuation_counts_for_order()` to decode varint keys
- Fix `compute_mkn_stats_single_trie_with_events()` to decode varint keys
- Add frequency count computation (n1, n2, n3, n4) to single-trie mode
- Start vocabulary indices at 1 (not 0) to avoid \x00 metadata collision
- Bump vocabulary version to 3 for the index offset change
- Add state machine for import workflow with cleanup guards
- Improve TUI with phase tracking and MKN progress display
- Add varint encoding benchmarks

+++ src/dictionary/vocabulary_backed.rs
@@ -86,7 +86,7 @@ impl<'a> VocabularyDictionary<'a> {
     }
 
     /// Get the vocabulary size (number of unique words).
-    pub fn len(&self) -> u32 {
+    pub fn len(&self) -> u64 {
         self.vocabulary.len()
     }
